Splitted nodearray.cpp for two files
